move server arguments
